aecp-aem: prepare field for in-coming changes
