+++ src/modules/meson.build
@@ -748,7 +748,8 @@ if build_module_avb
       'module-avb/msrp.c',
       'module-avb/mvrp.c',
       'module-avb/srp.c',
-      'module-avb/stream.c'
+      'module-avb/stream.c',
+      'module-avb/aecp-cmd-resp/aecp-aem-lock.c'
       ],
     include_directories : [configinc],
     install : true,
